+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the docs fetch mocking to now wait for the real request duration. This duration is obtained at build time from the manifold APIs. (#450)
 - Changed the `manifold-product` component to now use the GraphQL API rather than the REST API. (#456)
 - Changed `manifold-data-product-name` to use GraphQL (#463)
+- Changed `manifold-data-resource-list` to use GraphQL (#474)
 
 ## [v0.5.8]